requirements.txt
fastapi
sqlalchemy
asyncpg
xmltodict
python-jose[cryptography]
passlib[bcrypt]
pydantic
uvicorn

numpy
scipy
python-multipart
